Remove terraResourceTypeBI type
The terraResourceTypeBI type attempted to keep up with the number of
possible resource types, which has exploded recently.

Right now there are 70 different AWS types alone, with more being added.
The total count at the moment is 155.  Each provisioner adds more, so
it's impossible to keep up with.  Maybe we should just stop.

The @bkad version uses a Makefile to generate them, but they'll still go
stale with each release and it requires the TF code to rebuild.

Maintain the highlighting as Tag, but there's no keyword checking
anymore.  Spell carefully.